Fix cartridge overwriting
Addresses #7

It looks like the regex expects that there will always be a line break at the end of the cartridge. Since that is not the case (or maybe, no longer the case as of some newer version of PICO-8) the result is that it will find every section aside from the last section. This continues with each save until only the lua section is left. The fix is to alter the regex to allow for a line break OR the end of the file.

Tested with PICO-8 v0.1.11g for Mac.